Extract the OptionsPanel component
